requirements.txt
Pillow>=10.0.0
pillow-heif>=0.10.0
svglib>=1.5.1
reportlab>=4.0.0
